Rename search to searchInput and searchView to searchResult
