+++ app/components/search/searchResultView.js
